debugging halt

+++ include/vocab.h
@@ -19,7 +19,7 @@ extern std::unordered_map<std::string, int> vocab_hash;
 extern std::vector<word> vocab_list;
 extern std::string train_corpus_file, output_file;
 extern std::string save_vocab_file, read_vocab_file;
-extern size_t vocab_size;
+extern int vocab_size;
 extern long long train_words, file_size;
